chore(web): Separate component
