+++ web/src/beta/components/LayerSectionField/index.tsx
@@ -1,16 +1,8 @@
-import { useState } from "react";
-
 import Icon from "@reearth/beta/components/Icon";
 import SidePanelSectionField from "@reearth/beta/components/SidePanelSectionField";
-import Text from "@reearth/beta/components/Text";
 import { styled, useTheme } from "@reearth/services/theme";
 
-type LayerSectionItemProps = {
-  id: string;
-  label: string;
-  visible?: boolean;
-  active?: boolean;
-};
+import LayerSectionItem, { LayerSectionItemProps } from "./item";
 
 export type LayerSectionFieldProps = {
   layers: LayerSectionItemProps[];
@@ -28,7 +20,6 @@ const LayerSectionField: React.FC<LayerSectionFieldProps> = ({
   onVisible,
 }) => {
   const theme = useTheme();
-  const [visibles] = useState(layers);
 
   return (
     <SidePanelSectionField title="Layers">
@@ -42,35 +33,13 @@ const LayerSectionField: React.FC<LayerSectionFieldProps> = ({
       </LayerAddFrame>
       <List>
         {layers.map((item, index) => (
-          <Item
+          <LayerSectionItem
             key={index}
-            active={item.active}
-            onClick={() => {
-              onActive?.(item.id);
-            }}>
-            <ItemProperty>
-              <ItemVisibility
-                active={item.active}
-                color={theme.general.content.main}
-                onClick={() => onVisible?.(item.id)}>
-                {visibles[index].visible && <ItemVisibilityIcon icon="checkOnVisibility" />}
-              </ItemVisibility>
-              <Text
-                size="footnote"
-                color={theme.general.content.main}
-                otherProperties={{ height: "16px" }}>
-                {item.label}
-              </Text>
-            </ItemProperty>
-            <Icon
-              icon={"actionbutton"}
-              size={16}
-              color={theme.general.content.main}
-              onClick={() => {
-                onAction?.(item.id);
-              }}
-            />
-          </Item>
+            layer={item}
+            onActive={onActive}
+            onAction={onAction}
+            onVisible={onVisible}
+          />
         ))}
       </List>
     </SidePanelSectionField>
@@ -92,39 +61,6 @@ const List = styled.div`
   gap: 8px;
 
   width: 100%;
-`;
-const Item = styled.button<{ active?: boolean }>`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  gap: 4px;
-
-  align-self: stretch;
-
-  ${props => (props.active ? "background: " + props.theme.general.select + ";" : "")}
-  ${props => (props.active ? "border-radius: 4px;" : "")}
-  &:hover {${props =>
-    props.active ? "" : "background: " + props.theme.general.bg.weak + "; border-radius: 0px;"}
-`;
-const ItemProperty = styled.div`
-  display: flex;
-  align-items: center;
-  gap: 4px;
-`;
-const ItemVisibility = styled.div<{ active?: boolean; onClick?: () => void }>`
-  box-sizing: border-box;
-
-  width: 20px;
-  height: 20px;
-
-  border: 1px solid rgba(0, 0, 0, 0.25);
-  border-radius: 4px;
-`;
-const ItemVisibilityIcon = styled(Icon)`
-  padding-left: 30%;
-  padding-right: 33.44%;
-  padding-top: 30%;
-  padding-bottom: 27.34%;
 `;
 
 export default LayerSectionField;
